Fixes parsing problems when parsing implementation dependencies.
Sine the implementation path is checked against previously visited paths
and this is set before the implementation is parsed (immediately before)
the parsing of this implementation import is skipped.

We need to add a force flag because setting the implementation path
after the parsing may not be possible with the encapsulation the
function has. The function is passed a pointer to the file info
structure, yet, since the _done array does not store pointers to
FileInfos, this location may move.

I went for the quicker fix than the rightful one. I can keep the
abstraction without drastically changing the data structure.