Fixes ldc usage dump when attempting to compile when no implementation exists and quits on error.
Some djehuty interfaces, for instance, do not need implementations
because they are simply enums, etc. Therefore, they do not find
implementations. In this case, just have seshat skip compilation of
these implementations so that it does not invoke ldc with an empty
string for the source file.

Also, when errors occur, the compilation of the project stops.